+++ modules/server/variables.tf
@@ -26,12 +26,12 @@ variable "gsuite_admin_email" {
 
 variable "forseti_version" {
   description = "The version of Forseti to install"
-  default     = "v2.19.1"
+  default     = "v2.20.0"
 }
 
 variable "forseti_repo_url" {
   description = "Git repo for the Forseti installation"
-  default     = "https://github.com/GoogleCloudPlatform/forseti-security"
+  default     = "https://github.com/forseti-security/forseti-security"
 }
 
 variable "forseti_email_recipient" {
@@ -583,17 +583,17 @@ variable "inventory_email_summary_enabled" {
 # Forseti server #
 #----------------#
 variable "server_type" {
-  description = "GCE Forseti Server role instance size"
+  description = "GCE Forseti Server machine type"
   default     = "n1-standard-2"
 }
 
 variable "server_region" {
-  description = "GCP region where Forseti will be deployed"
+  description = "GCE Forseti Server region"
   default     = "us-central1"
 }
 
 variable "server_boot_image" {
-  description = "GCE instance image that is being used, currently Debian only support is available"
+  description = "GCE Forseti Server boot image - Currently only Ubuntu is supported"
   default     = "ubuntu-os-cloud/ubuntu-1804-lts"
 }
 
@@ -632,39 +632,6 @@ variable "server_private" {
 
 variable "client_service_account_email" {
   description = "Service account of the forseti client"
-}
-
-#------------#
-# Forseti db #
-#------------#
-variable "cloudsql_region" {
-  description = "CloudSQL region"
-  default     = "us-central1"
-}
-
-variable "cloudsql_db_name" {
-  description = "CloudSQL database name"
-  default     = "forseti_security"
-}
-
-variable "cloudsql_db_port" {
-  description = "CloudSQL database port"
-  default     = "3306"
-}
-
-variable "cloudsql_proxy_arch" {
-  description = "CloudSQL Proxy architecture"
-  default     = "linux.amd64"
-}
-
-variable "cloudsql_type" {
-  description = "CloudSQL Instance size"
-  default     = "db-n1-standard-1"
-}
-
-variable "cloudsql_user_host" {
-  description = "The host the user can connect from. Can be an IP address or IP address range. Changing this forces a new resource to be created."
-  default     = "%"
 }
 
 #----------------#
@@ -763,6 +730,19 @@ variable "services" {
   description = "An artificial dependency to bypass #10462"
   type        = list(string)
   default     = []
+}
+
+#------------#
+# Forseti db #
+#------------#
+
+variable "cloudsql_module" {
+  description = "The CloudSQL module"
+}
+
+variable "cloudsql_proxy_arch" {
+  description = "CloudSQL Proxy architecture"
+  default     = "linux.amd64"
 }
 
 #---------------------------------------#