+++ CHANGELOG.md
@@ -15,6 +15,15 @@ Extending the adopted spec, each change should have a link to its corresponding 
   - A variable to toggle management of rules
   - An import helper script
 
+- CloudSQL instance created in the same zone as GCE instances [#253]
+
+## [v4.1.0] - 2019-09-06
+
+### Added
+
+- Support for Forseti v2.20.0 [#246]
+- Added support for Policy Library sync. [#239]
+
 ## [v4.0.1] - 2019-08-23
 
 ### Added
@@ -245,7 +254,11 @@ Version 4.0.0 is a backwards-incompatible release. Please see the [upgrade instr
 [v3.0.0]: https://github.com/terraform-google-modules/terraform-google-forseti/compare/v2.3.0...v3.0.0
 [v4.0.0]: https://github.com/terraform-google-modules/terraform-google-forseti/compare/v3.0.0...v4.0.0
 [v4.0.1]: https://github.com/terraform-google-modules/terraform-google-forseti/compare/v4.0.0...v4.0.1
+[v4.1.0]: https://github.com/terraform-google-modules/terraform-google-forseti/compare/v4.0.1...v4.1.0
 
+[#253]: https://github.com/forseti-security/terraform-google-forseti/pull/253
+[#246]: https://github.com/forseti-security/terraform-google-forseti/pull/246
+[#239]: https://github.com/forseti-security/terraform-google-forseti/pull/239
 [#233]: https://github.com/forseti-security/terraform-google-forseti/pull/233
 [#232]: https://github.com/forseti-security/terraform-google-forseti/pull/232
 [#231]: https://github.com/forseti-security/terraform-google-forseti/pull/231